Day 1

+++ database/migrations/2019_02_16_211958_create_products_table.php
@@ -18,7 +18,7 @@ class CreateProductsTable extends Migration
            $table->integer('category_id')->unsigned();
             $table->integer('genderId');
             $table->string('name');
-            $table->integer('price');
+            $table->float('price');
             $table->string('size');
             $table->binary('pic');
             $table->timestamps();
